split top level to operation specific modules.
