+++ raaz/Raaz/V1.hs
@@ -5,16 +5,9 @@
 --
 -- For documentation refer the top-most module "Raaz".
 
-module Raaz.V1 ( Digest
-               , Auth
-               , module Digest.Blake2b
-               , module Auth.Blake2b
+module Raaz.V1 ( module Raaz.V1.Digest
+               , module Raaz.V1.Auth
                ) where
 
-import Digest.Blake2b
-import Auth.Blake2b
-import Raaz.Primitive.Blake2.Internal(Blake2b)
-import Raaz.Primitive.Keyed.Internal(Keyed)
-
-type Digest = Blake2b
-type Auth   = Keyed Blake2b
+import Raaz.V1.Digest
+import Raaz.V1.Auth
